Extinguishing fire and recruiting animals

+++ rules/src/rules/PlayerActionRule.ts
@@ -4,9 +4,19 @@ import { LocationType } from '../material/LocationType'
 import { Animal, animalProperties, AnimalSeason, AnimalType, CardElements, getAnimalSeason } from '../material/Animal'
 import { AnimalsHelper } from './helpers/AnimalsHelper'
 import { countBy } from 'lodash'
-import { seasons } from '../Season'
+import { Element, seasons } from '../Season'
+import { RuleId } from './RuleId'
+import { Memory } from './Memory'
 
 export class PlayerActionRule extends PlayerTurnRule {
+  onRuleStart() {
+    if (this.material(MaterialType.ActionToken).location(LocationType.PlayerActionSupply).id(this.player).getQuantity() === 0) {
+      return [this.startRule(RuleId.CheckEndTurn)]
+    } else {
+      return []
+    }
+  }
+
   getPlayerMoves() {
     const moves: MaterialMove[] = []
 
@@ -14,7 +24,8 @@ export class PlayerActionRule extends PlayerTurnRule {
     const sharedDeck = this.material(MaterialType.AnimalCard).location(LocationType.SharedDeck).deck()
     moves.push(sharedDeck.dealOne({ type: LocationType.SharedHelpLine }))
     // Or add one action token to one card in the shared help line
-    if (this.availableActionTokens.getQuantity() > 0) {
+    if (this.availableActionTokens.getQuantity() > 0
+        && this.getActivePlayer() === this.game.players[0]) { // TODO: Remove this. Just for testing purposes to avoid monkey opponent playing the action
       moves.push(...this.getAvailableActions())
     }
 
@@ -75,26 +86,50 @@ export class PlayerActionRule extends PlayerTurnRule {
     const moves: MaterialMove[] = []
     if (isMoveItemType(MaterialType.AnimalCard)(move) && move.location.type !== LocationType.PersonalHelpLine) {
       moves.push(...this.drawCard(move))
-      // const movedAnimal = this.material(MaterialType.AnimalCard).getItem(move.itemIndex)
-      // if (getAnimalSeason(movedAnimal.id) !== AnimalSeason.Common && movedAnimal.id !== Animal.Stag) {
-      //   const animalSeason = getAnimalSeason(movedAnimal.id)
-      //   moves.push(this.material(MaterialType.AnimalCard).id(movedAnimal.id).moveItem({ type: LocationType.PersonalHelpLine, id: animalSeason }))
-      // }
-      // const movedAnimalProperties = animalProperties[movedAnimal.id as Animal]
-      // if (movedAnimalProperties?.type === AnimalType.Solitary) {
-      //   // Check number of solitary symbols
-      //   for (const season of seasons) {
-      //     if (this.checkTooManySolitaryAnimals(season)) {
-      //       // TODO: Implement this in other way (e.g. disabling or with a cross) or fade it out
-      //       moves.push(this.material(MaterialType.ActionToken).location(LocationType.PlayerActionSupply).id(season).deleteItem())
-      //     }
-      //   }
-      // }
-    } else if (isMoveItemType(MaterialType.ActionToken)(move)) {
-      moves.push(...this.useActionToken(move))
+      moves.push(this.startRule(RuleId.CheckEndTurn))
+    } else if (isMoveItemType(MaterialType.ActionToken)(move) && move.location.type === LocationType.ActionToken) {
+      switch(move.location.y) {
+        case Element.Sun:
+          this.setRemainingElementValue(Element.Sun)
+          moves.push(this.startRule(RuleId.RecruitingAnimals))
+          break
+        case Element.Water:
+          this.setRemainingElementValue(Element.Water)
+          moves.push(this.startRule(RuleId.ExtinguishingFire))
+          break
+        case Element.Plant:
+          moves.push(this.startRule(RuleId.PlantingProtectiveTree))
+          break
+        case Element.Wind:
+          moves.push(this.startRule(RuleId.AdvancingOnibi))
+          break
+      }
     }
-
+    
+    
     return moves
+  }
+
+  setRemainingElementValue(elementType: Element) {
+    let elementValue = 0
+
+    const tokensLocations = this.material(MaterialType.ActionToken)
+      .location(l => l.type === LocationType.ActionToken && l.y === elementType)
+      .getItems()
+      .sort((a,b) => b.location.x! - a.location.x!)
+    const tokenLocationX = tokensLocations[0].location.x
+    const previousTokenLocationX = tokensLocations[1]?.location.x ?? -1
+    for (let x = tokenLocationX!; x > previousTokenLocationX; x--) {
+      const card = this.material(MaterialType.AnimalCard).location(l => l.type === LocationType.SharedHelpLine && l.x === x).getItem()
+      const cardProperties = animalProperties[card?.id as Animal]
+      elementValue += cardProperties?.elements[Element[elementType].toLowerCase() as keyof CardElements]! ?? 0
+    }
+    // Add the personal value
+    const playerCardsids = this.material(MaterialType.AnimalCard).location(l => l.type === LocationType.PersonalHelpLine && l.id === this.player).getItems().map(card => card.id)
+    elementValue += new AnimalsHelper(this.game, this.player).getAnimalsCostSum(playerCardsids)
+
+    console.log("Computed element value: ", elementValue)
+    this.memorize(Memory.RemainingElementValue, elementValue)
   }
 
   drawCard(move: ItemMove<number, number, number>) {
@@ -110,8 +145,11 @@ export class PlayerActionRule extends PlayerTurnRule {
         // Check number of solitary symbols
         for (const season of seasons) {
           if (this.checkTooManySolitaryAnimals(season)) {
-            // TODO: Implement this in other way (e.g. disabling or with a cross) or fade it out
-            moves.push(this.material(MaterialType.ActionToken).location(LocationType.PlayerActionSupply).id(season).deleteItem())
+            // TODO: Remove this "if" when the rules workflow is correctly implemented as it should not be necessary
+            if (this.material(MaterialType.ActionToken).location(LocationType.PlayerActionSupply).id(season).getQuantity() > 0) {
+              // TODO: Implement this in other way (e.g. disabling or with a cross) or fade it out
+              moves.push(this.material(MaterialType.ActionToken).location(LocationType.PlayerActionSupply).id(season).deleteItem())
+            }
           }
         }
       }
@@ -134,10 +172,4 @@ export class PlayerActionRule extends PlayerTurnRule {
 
     return false
   }
-
-  useActionToken(_move: ItemMove<number, number, number>) {
-    const moves: MaterialMove[] = []
-    return moves
-  }
-
 }
